fix(contention): address clippy cast_sign_loss in threshold calculation
Replace direct `as f64` cast of u64 with `f64::from(base_contentions as
u32)` to avoid potential sign-loss warning, and add explicit
`#[allow(clippy::cast_sign_loss)]` for the ceil-to-u64 conversion.

+++ tests/contention_e2e.rs
@@ -1053,7 +1053,8 @@ fn contention_post_sharding_comparison() {
         let test_name = base["test"].as_str().unwrap_or("?");
 
         // Allow up to 30% variance (timing-dependent measurements).
-        let threshold = (base_contentions as f64 * 1.3).ceil() as u64 + 10;
+        #[allow(clippy::cast_sign_loss)]
+        let threshold = (f64::from(base_contentions as u32) * 1.3).ceil() as u64 + 10;
         assert!(
             mono_contentions <= threshold,
             "Regression in {test_name}: contentions {mono_contentions} exceeds baseline {base_contentions} + 30% margin ({threshold})"
